feat: Factor out local date time helper, add types, include clarifying comment datetime
