+++ src/utils/localizeDateHelper.ts
@@ -1,0 +1,43 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const tz = require('timezone')(require('timezone/zones'), require('timezone/en_US.js'), require('timezone/sv_SE.js'), require('timezone/fr_FR.js'), require('timezone/nb_NO.js'), require('timezone/de_DE'), require('timezone/es_ES'), require('timezone/ja_JP'), require('timezone/ko_KR'));
+
+// tz docs https://bigeasy.github.io/timezone/
+function localizeDateHelper(
+  date: string,
+  targetDateFormat: string,
+  language: string,
+  timeZone: string,
+): string {
+  let locale = null;
+  switch (language) {
+    case 'sv':
+      locale = 'sv_SE';
+      break;
+    case 'fr':
+      locale = 'fr_FR';
+      break;
+    case 'no':
+      locale = 'nb_NO';
+      break;
+    case 'de':
+      locale = 'de_DE';
+      break;
+    case 'es':
+      locale = 'es_ES';
+      break;
+    case 'ja':
+      locale = 'ja_JP';
+      break;
+    case 'ko':
+      locale = 'ko_KR';
+      break;
+    case 'en':
+    default:
+      locale = 'en_US';
+  }
+  // Convert to UTC date
+  const utc = tz(date);
+  return tz(utc, targetDateFormat, locale, timeZone);
+}
+
+export default localizeDateHelper;
